requirements.txt
Werkzeug==2.2.2
Flask==2.0.3
pycairo
pydantic